requirements.txt
pyfiglet
serial
pycrypto
pyOSC
